Get rid of the concept of self.app.
From this point on, Flask-Bootstrap should be an extension that can
safely be used by multiple apps. Configuration values are looked up at
runtime or when init_app is called on the app itself. Referencing the
app is made solely by calling current_app.

For now, there really isn't any reason why a Flask-Bootstrap extension
instance should be specific to any app, as all the configuration is
stored on the app itself.